requirements.txt
boto3==1.21.21
click==8.0.4
flake8==4.0.1
kafka-python==2.0.2
pytest==7.1.1
python-dotenv==0.19.2
redis==4.1.4
requests==2.27.1
celery==5.2.3
udata_event_service==0.0.6
